# Conflicts:
#	app/src/main/java/eu/wise_iot/wanderlust/controllers/TourController.java
#	app/src/main/java/eu/wise_iot/wanderlust/models/DatabaseModel/UserTour.java
#	app/src/main/java/eu/wise_iot/wanderlust/services/ServiceGenerator.java
#	app/src/main/java/eu/wise_iot/wanderlust/views/TourFragment.java
#	app/src/main/java/eu/wise_iot/wanderlust/views/TourOverviewFragment.java
#	app/src/main/java/eu/wise_iot/wanderlust/views/adapters/MyRecyclerViewAdapter.java

+++ app/src/main/java/eu/wise_iot/wanderlust/services/UserTourService.java
@@ -31,7 +31,7 @@ import retrofit2.http.Path;
  */
 public interface UserTourService {
     @GET("tour/{id}")
-    Call<UserTour> retrieveUserTour(@Path("id") int id);
+    Call<UserTour> retrieveUserTour(@Path("id") long id);
 
     @GET("tour/")
     Call<List<UserTour>> retrieveAllUserTours();
